refactor(trace): share fixed-width ID encoding
Keep the agentless and JSON log encoders on one stack-backed implementation so their fixed-width hexadecimal wire formats cannot drift.
