+++ libdd-trace-utils/src/agentless_encoder/mod.rs
@@ -26,14 +26,13 @@
 //!
 //! TODO: span normalization (service/name/resource/type truncation + defaults)
 
+use crate::hex::{hex_low_u64, hex_u128, hex_u64};
 use crate::span::v04::{AttributeAnyValue, AttributeArrayValue, Span, SpanEvent, SpanLink};
 use crate::span::v1;
 use crate::span::{TraceData, SPAN_LINK_FLAGS_SET_SENTINEL};
 use crate::tracer_metadata::TracerMetadata;
-use serde::{
-    ser::{Error as _, SerializeMap, SerializeSeq},
-    Serializer,
-};
+use serde::ser::{SerializeMap, SerializeSeq};
+use serde::Serializer;
 use std::borrow::{Borrow, Cow};
 use std::collections::HashSet;
 use std::fmt::Write as _;
@@ -42,31 +41,6 @@ use std::fmt::Write as _;
 const MAX_META_VALUE_LEN: usize = 25_000;
 /// Suffix appended when a `meta` value is truncated.
 const TRUNCATION_SUFFIX: &str = "...";
-const HEX_DIGITS: &[u8; 16] = b"0123456789abcdef";
-
-struct FixedHex<const N: usize>([u8; N]);
-
-impl<const N: usize> serde::Serialize for FixedHex<N> {
-    fn serialize<S: Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
-        let value = std::str::from_utf8(&self.0).map_err(S::Error::custom)?;
-        serializer.serialize_str(value)
-    }
-}
-
-fn fixed_hex<const N: usize>(mut value: u128) -> FixedHex<N> {
-    let mut encoded = [b'0'; N];
-    let mut index = N;
-
-    while index != 0 {
-        index -= 1;
-        // The mask guarantees that the conversion fits and indexes HEX_DIGITS.
-        encoded[index] = HEX_DIGITS[(value & 0x0f) as usize];
-        value >>= 4;
-    }
-
-    FixedHex(encoded)
-}
-
 /// # Why are we doing this?
 ///
 /// The JSON agentless format is different from the in-memory model of v04 spans (and to v03/v04/v05
@@ -199,9 +173,9 @@ fn encode_span<T: TraceData, S: Serializer>(
 ) -> Result<S::Ok, S::Error> {
     let mut map = ser.serialize_map(None)?;
 
-    map.serialize_entry("trace_id", &fixed_hex::<16>(span.trace_id))?;
-    map.serialize_entry("span_id", &fixed_hex::<16>(u128::from(span.span_id)))?;
-    map.serialize_entry("parent_id", &fixed_hex::<16>(u128::from(span.parent_id)))?;
+    map.serialize_entry("trace_id", &hex_low_u64(span.trace_id))?;
+    map.serialize_entry("span_id", &hex_u64(span.span_id))?;
+    map.serialize_entry("parent_id", &hex_u64(span.parent_id))?;
 
     // Resource defaults to name when empty.
     let name_str: &str = span.name.borrow();
@@ -249,7 +223,7 @@ fn encode_span<T: TraceData, S: Serializer>(
                 meta.serialize_entry(key, val)?;
             }
             if !p_tid_seen && upper_bits != 0 {
-                meta.serialize_entry("_dd.p.tid", &fixed_hex::<16>(u128::from(upper_bits)))?;
+                meta.serialize_entry("_dd.p.tid", &hex_u64(upper_bits))?;
             }
             if !span_links_seen && !span.span_links.is_empty() {
                 if let Some(s) = serialize_span_links(&span.span_links) {
@@ -339,8 +313,8 @@ fn encode_span_link<T: TraceData, S: Serializer>(
 ) -> Result<S::Ok, S::Error> {
     let mut map = ser.serialize_map(None)?;
     let trace_id_128 = (u128::from(link.trace_id_high) << 64) | u128::from(link.trace_id);
-    map.serialize_entry("trace_id", &fixed_hex::<32>(trace_id_128))?;
-    map.serialize_entry("span_id", &fixed_hex::<16>(u128::from(link.span_id)))?;
+    map.serialize_entry("trace_id", &hex_u128(trace_id_128))?;
+    map.serialize_entry("span_id", &hex_u64(link.span_id))?;
     if !link.attributes.is_empty() {
         map.serialize_entry(
             "attributes",
@@ -720,22 +694,9 @@ fn encode_span_v1<'a, T: TraceData, S: Serializer>(
     trace_id_high_bytes.copy_from_slice(&chunk.trace_id[0..8]);
     let trace_id_low = u64::from_be_bytes(trace_id_low_bytes);
     let trace_id_high = u64::from_be_bytes(trace_id_high_bytes);
-    map.serialize_entry(
-        "trace_id",
-        &ser_fn!(|ser, trace_id_low: u64| {
-            ser.collect_str(&format_args!("{trace_id_low:016x}"))
-        }),
-    )?;
-    let span_id = span.span_id;
-    map.serialize_entry(
-        "span_id",
-        &ser_fn!(|ser, span_id: u64| { ser.collect_str(&format_args!("{span_id:016x}")) }),
-    )?;
-    let parent_id = span.parent_id;
-    map.serialize_entry(
-        "parent_id",
-        &ser_fn!(|ser, parent_id: u64| { ser.collect_str(&format_args!("{parent_id:016x}")) }),
-    )?;
+    map.serialize_entry("trace_id", &hex_u64(trace_id_low))?;
+    map.serialize_entry("span_id", &hex_u64(span.span_id))?;
+    map.serialize_entry("parent_id", &hex_u64(span.parent_id))?;
 
     // Resource defaults to name when empty.
     let name_str: &str = span.name.borrow();
@@ -818,12 +779,7 @@ fn encode_span_v1<'a, T: TraceData, S: Serializer>(
                 meta.serialize_entry(key, val)?;
             }
             if !p_tid_seen && trace_id_high != 0 {
-                meta.serialize_entry(
-                    "_dd.p.tid",
-                    &ser_fn!(|ser, trace_id_high: u64| {
-                        ser.collect_str(&format_args!("{trace_id_high:016x}"))
-                    }),
-                )?;
+                meta.serialize_entry("_dd.p.tid", &hex_u64(trace_id_high))?;
             }
             if !span_links_seen && !span.span_links.is_empty() {
                 if let Some(s) = serialize_span_links_v1(&span.span_links) {
@@ -913,8 +869,8 @@ fn encode_span_link_v1<T: TraceData, S: Serializer>(
 ) -> Result<S::Ok, S::Error> {
     let mut map = ser.serialize_map(None)?;
     let trace_id_128 = u128::from_be_bytes(link.trace_id);
-    map.serialize_entry("trace_id", &format!("{trace_id_128:032x}"))?;
-    map.serialize_entry("span_id", &format!("{:016x}", link.span_id))?;
+    map.serialize_entry("trace_id", &hex_u128(trace_id_128))?;
+    map.serialize_entry("span_id", &hex_u64(link.span_id))?;
     let attrs_dd = link.attributes.defensive_dedup();
     let attrs_dd = &attrs_dd;
     let has_attributes = attrs_dd.iter().any(|(_, v)| {
